refactored CurrentSongView
